:construction: Bootloader and Host application CAN driver updated. 1. Host application CAN driver SEG fault fixed.

+++ Bootloader/CAN_UART_Bootloader/CAN_UART_Bootloader/OTC_Sources/data_type_support.h
@@ -43,9 +43,9 @@ enum states{
 /* Main application to can driver support */
 typedef struct _can_context //
 {
-	uint8_t can_id;
+	uint32_t can_id;
 	uint8_t can_data[64];
-	uint8_t len;
+	uint32_t len;
 	char * can_serial_port;
 }can_context_type;
